dedicated file for diatomic utils
